Calculations in long

+++ src/Main.java
@@ -1,9 +1,9 @@
 public class Main {
 
     public static void main(String[] args) {
-        int [] sales = new int[] {15,0,25,10,38};
+        long [] sales = new long[] {15,0,25,10,38};
         SalesManager max = new SalesManager(sales);
-        int maxFin = max.max();
+        long maxFin = max.max();
         System.out.println("Максимальное количество продаж составляет: "+maxFin);
     }
 }
